NODE-312: Moved cats effect instances to node package.
